backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Phillipml",email = "contato.phillip.menezes@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=7.0.0,<8.0.0)",
    "django-debug-toolbar (>=6.0.0,<7.0.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "pillow (>=12.0.0,<13.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-django = "^4.11.1"
pytest-cov = "^7.0.0"
black = "^25.9.0"
flake8 = "^7.3.0"
bandit = "^1.8.6"
pre-commit = "^4.3.0"
mypy = "^1.18.2"
django-stubs = "^5.2.7"
isort = "^7.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
